+++ packages/protocol/contracts/layer1/based/libs/LibVerify.sol
@@ -80,12 +80,12 @@ library LibVerify {
 
                     {
                         uint96 bondToReturn;
-                        if (ts.proofTiming == uint8(ITaikoInbox.ProofTiming.InProvingWindow)) {
+                        if (ts.proofTiming == ITaikoInbox.ProofTiming.InProvingWindow) {
                             // all liveness bond is returned to the prover, this is not a reward.
                             bondToReturn = batch.livenessBond;
                             if (tid == 1) bondToReturn += batch.provabilityBond;
                         } else if (
-                            ts.proofTiming == uint8(ITaikoInbox.ProofTiming.InExtendedProvingWindow)
+                            ts.proofTiming == ITaikoInbox.ProofTiming.InExtendedProvingWindow
                         ) {
                             // prover is rewarded with bondRewardPtcg% of the liveness bond.
                             bondToReturn = batch.livenessBond * _config.bondRewardPtcg / 100;
